reorganizes theme
