+++ client/stories/index.stories.js
@@ -1,7 +1,7 @@
 import { configure, addDecorator } from "@storybook/react";
 import { ThemeProvider } from "emotion-theming";
 import React from "react";
-import theme from "../uikit/defaultTheme";
+import theme from "../uikit/theme/defaultTheme";
 
 const req = require.context("../uikit", true, /.stories\.js$/);
 function loadStories() {
