+++ client/uikit/theme/defaultTheme/index.js
@@ -1,0 +1,7 @@
+import colors from "./colors";
+import button from "./button";
+
+export default {
+  colors,
+  button
+};
