+++ client/uikit/utils/useTheme.js
@@ -1,0 +1,3 @@
+import { ThemeContext } from "@emotion/core";
+import { useContext } from "react";
+export default () => useContext(ThemeContext);
